feat: アップロードとアーカイブ全体で一貫したパス生成を実現するため、slugifyForPathユーティリティを追加する
