+++ app/routes/api.onedrive.upload.tsx
@@ -27,6 +27,7 @@ import { verifyCsrfToken } from "../services/csrf.server";
 import { normalizeEvidenceSourceUrl } from "../services/evidence-url";
 import { validatePrRefInput } from "../services/validation";
 import { signEvidenceImagePath } from "../services/evidence-image-token.server";
+import { slugifyForPath } from "../services/path-utils";
 
 // ルートハンドラーとビジネスロジックを分離するため、OneDrive への保存処理の詳細は services/evidence-images.server.ts に委譲する。
 export type ApiOneDriveUploadResponse =
@@ -501,38 +502,6 @@ function formatIsoForJst(date: Date): string {
   const offsetHours = String(Math.floor(abs / 60)).padStart(2, "0");
   const offsetMins = String(abs % 60).padStart(2, "0");
   return `${year}-${month}-${day}T${hours}:${minutes}:${seconds}${sign}${offsetHours}:${offsetMins}`;
-}
-
-function slugifyForPath(value: string): string {
-  const normalized = value
-    .normalize("NFC")
-    // OneDrive/Windowsで禁止される文字だけ除去し、日本語は保持する。
-    .replace(/[<>:"/\\|?*\u0000-\u001F]/g, "")
-    .trim()
-    .replace(/\s+/g, "-")
-    .replace(/-+/g, "-")
-    .replace(/^[-.\s]+|[-.\s]+$/g, "");
-
-  // サロゲートペアを壊さず、かつ過度に長いマルチバイト文字列を避けるため、
-  // コードポイント数と UTF-8 バイト長の両方で上限を適用する。
-  const maxCodePoints = 80;
-  const maxUtf8Bytes = 160;
-  const encoder = new TextEncoder();
-  let result = "";
-  let codePointCount = 0;
-  let utf8ByteCount = 0;
-
-  for (const char of normalized) {
-    const charBytes = encoder.encode(char).length;
-    if (codePointCount + 1 > maxCodePoints || utf8ByteCount + charBytes > maxUtf8Bytes) {
-      break;
-    }
-    result += char;
-    codePointCount += 1;
-    utf8ByteCount += charBytes;
-  }
-
-  return result;
 }
 
 /*
